@@ -1,86 +1,84 @@
 {
-  "name": "xpx-catapult-web-wallet",
-  "version": "0.1.6",
+  "name": "proximax-sirius-wallet",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "main": "xpx-catapult-web-wallet/dist/xpx-catapult-web-wallet/main.js",
+    "main": "proximax-sirius-wallet/dist/proximax-sirius-wallet/main.js",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "electron": "electron .",
     "electron-build": "ng build && electron .",
-    "package-mac": "electron-packager . xpx-catapult-web-wallet --overwrite --asar=true --platform=darwin --arch=x64 --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1",
-    "package-linux": "electron-packager . xpx-catapult-web-wallet --overwrite --asar=true --platform=linux --arch=x64 --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1",
-    "package-win": "electron-packager . xpx-catapult-web-wallet --overwrite --asar=true --platform=win32 --arch=x64  --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1"
+    "package-mac": "electron-packager . proximax-sirius-wallet --overwrite --asar=true --platform=darwin --arch=x64 --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1",
+    "package-linux": "electron-packager . proximax-sirius-wallet --overwrite --asar=true --platform=linux --arch=x64 --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1",
+    "package-win": "electron-packager . proximax-sirius-wallet --overwrite --asar=true --platform=win32 --arch=x64  --prune=true --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName='ProximaX Web Wallet' --ignore=node_modules --ignore=src --ignore=e2e --ignore=release-buildsv1 --out=release-buildsv1",
+    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+    "serve:ssr": "node dist/server",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "build:client-and-server-bundles": "ng build --prod && ng run proximax-sirius-wallet:server:production --bundleDependencies all"
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@ng-select/ng-select": "^2.20.0",
-    "Buffer": "0.0.0",
-    "ajv": "^5.5.2",
+    "@agm/core": "^1.0.0-beta.7",
+    "@angular/animations": "~8.0.1",
+    "@angular/common": "~8.0.1",
+    "@angular/compiler": "~8.0.1",
+    "@angular/core": "~8.0.1",
+    "@angular/forms": "~8.0.1",
+    "@angular/platform-browser": "~8.0.1",
+    "@angular/platform-browser-dynamic": "~8.0.1",
+    "@angular/platform-server": "~8.0.1",
+    "@angular/router": "~8.0.1",
+    "@ng-select/ng-select": "^3.0.3",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
     "bootstrap": "^4.3.1",
     "chart.js": "^2.8.0",
-    "core-js": "^2.5.4",
-    "crypto-js": "^3.1.9-1",
     "easy-pie-chart": "^2.1.7",
+    "electron": "^5.0.7",
+    "electron-builder": "^21.1.1",
+    "express": "^4.15.2",
     "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "highcharts": "^7.1.1",
-    "js-joda": "^1.10.1",
     "js-xpx-chain-library": "^0.3.0",
-    "jspdf": "^1.5.3",
-    "moment": "^2.24.0",
-    "ng-block-ui": "^2.1.3",
-    "ng-uikit-pro-standard": "git+https://oauth2:xds7fpVXkw8SnNJdDtUE@git.mdbootstrap.com/mdb/angular/ng-uikit-pro-standard.git#7.5.4",
-    "ngx-clipboard": "^12.0.0",
+    "ng-block-ui": "^2.1.7",
+    "ng-uikit-pro-standard": "git+https://oauth2:xds7fpVXkw8SnNJdDtUE@git.mdbootstrap.com/mdb/angular/ng-uikit-pro-standard.git",
+    "ngx-clipboard": "^12.2.0",
     "ngx-infinite-scroll": "^7.2.0",
-    "ngx-mask": "^7.9.10",
-    "ngx-numeral": "0.0.1-beta.0",
-    "ngx-pagination": "^4.0.0",
-    "qrcode-generator": "^1.4.3",
-    "rxjs": "^6.5.2",
+    "ngx-mask": "^8.0.2",
+    "ngx-pagination": "^4.1.0",
+    "rxjs": "~6.4.0",
     "screenfull": "^4.2.0",
     "tsjs-xpx-chain-sdk": "^0.2.0",
     "tslib": "^1.9.0",
-    "utf8": "^2.1.2",
-    "xpx2-ts-js-sdk": "0.0.20",
-    "zone.js": "~0.8.26"
+    "xpx2-ts-js-sdk": "^0.0.51",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/crypto-js": "^3.1.43",
-    "@types/electron": "^1.6.10",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler-cli": "~8.0.1",
+    "@angular/language-service": "~8.0.1",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.10.51",
-    "codelyzer": "~4.5.0",
-    "electron": "^5.0.7",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
+    "ts-loader": "^5.2.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3",
+    "webpack-cli": "^3.1.0"
   },
   "browser": {
     "fs": false,
