@@ -1,14 +1,13 @@
 import { Component } from '@angular/core';
-import { NodeService } from "./servicesModule/services/node.service";
+import { NodeService } from './servicesModule/services/node.service';
 
 @Component({
   selector: 'app-root',
-  template: `<block-ui>
-                <app-header></app-header>
-                <router-outlet></router-outlet>
-              </block-ui>`
+  template: `<router-outlet></router-outlet>`
 })
+
 export class AppComponent {
+  title = 'Sirius Wallet';
 
   constructor(private nodeService: NodeService) {
     this.nodeService.initNode();
