@@ -2,15 +2,18 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AuthComponent } from './views/auth/auth.component';
 import { AppConfig } from '../config/app.config';
-import { NotLoggedGuard } from '../shared/guard/not-logged.guard';
 
-const routes: Routes = [
-
-  {   path: `${AppConfig.routes.login}`,
-      component: AuthComponent,
-      canActivate: [NotLoggedGuard]
-  }
-];
+const routes: Routes = [{
+  path: AppConfig.routes.auth,
+  component: AuthComponent,
+  data: {
+    meta: {
+      title: 'auth.title',
+      description: 'auth.text',
+      override: true,
+    },
+  },
+}];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
