@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-lock-type',
   templateUrl: './lock-type.component.html',
-  styleUrls: ['./lock-type.component.scss']
+  styleUrls: ['./lock-type.component.css']
 })
 export class LockTypeComponent implements OnInit {
 
