@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-modify-multisig-account-type',
   templateUrl: './modify-multisig-account-type.component.html',
-  styleUrls: ['./modify-multisig-account-type.component.scss']
+  styleUrls: ['./modify-multisig-account-type.component.css']
 })
 export class ModifyMultisigAccountTypeComponent implements OnInit {
 
