@@ -1,15 +1,13 @@
 import { Component, OnInit, Input, SimpleChanges } from '@angular/core';
 import { MosaicId } from 'tsjs-xpx-chain-sdk';
 import * as cloneDeep from 'lodash/cloneDeep';
-import { TransactionsService } from '../../../transactions/service/transactions.service';
-import { TransactionsInterface } from '../../services/transaction.interface';
-import { MosaicService } from '../../../servicesModule/services/mosaic.service';
-import { MosaicsStorage } from '../../../servicesModule/interfaces/mosaics-namespaces.interface';
+import { MosaicService, MosaicsStorage } from '../../../servicesModule/services/mosaic.service';
+import { TransactionsService, TransactionsInterface } from '../../../transfer/services/transactions.service';
 
 @Component({
   selector: 'app-mosaic-definition-type',
   templateUrl: './mosaic-definition-type.component.html',
-  styleUrls: ['./mosaic-definition-type.component.scss']
+  styleUrls: ['./mosaic-definition-type.component.css']
 })
 export class MosaicDefinitionTypeComponent implements OnInit {
 
