@@ -1,21 +1,19 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
-import { AppConfig } from '../config/app.config';
-import { LoggedGuard } from '../shared/guard/logged.guard';
 import { DashboardComponent } from './views/dashboard/dashboard.component';
-
+import { AppConfig } from '../config/app.config';
 
 const routes: Routes = [
-  // {
-  //   path: `${AppConfig.routes.dashboard}`,
-  //   component: DashboardComponent2,
-  //   canActivate: [LoggedGuard]
-  // },
   {
-    path: `${AppConfig.routes.dashboard}`,
+    path: AppConfig.routes.dashboard,
     component: DashboardComponent,
-    canActivate: [LoggedGuard]
+    data: {
+      meta: {
+        title: 'dashboard.title',
+        description: 'dashboard.text',
+        override: true,
+      },
+    }
   }
 ];
 
@@ -23,6 +21,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class DashboardRoutingModule {
-
-}
+export class DashboardRoutingModule { }
