@@ -1,14 +1,19 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './views/home/home.component';
 import { AppConfig } from '../config/app.config';
-import { NotLoggedGuard } from "../shared";
+import { HomeComponent } from './views/home/home.component';
 
 const routes: Routes = [
   {
-    path: `${AppConfig.routes.home}`,
+    path: AppConfig.routes.home,
     component: HomeComponent,
-    canActivate: [NotLoggedGuard]
+    data: {
+      meta: {
+        title: 'home.title',
+        description: 'home.text',
+        override: true,
+      },
+    }
   }
 ];
 
