@@ -1,13 +1,16 @@
 import { NgModule } from '@angular/core';
+
 import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './views/home/home.component';
 import { CoreModule } from '../core/core.module';
 
 @NgModule({
+  declarations: [
+    HomeComponent
+  ],
   imports: [
     CoreModule,
     HomeRoutingModule
-  ],
-  declarations: [HomeComponent]
+  ]
 })
 export class HomeModule { }
