@@ -2,12 +2,11 @@ import { Injectable } from '@angular/core';
 import { first } from "rxjs/operators";
 import { BehaviorSubject, Observable } from 'rxjs';
 import { Listener, Transaction, TransactionStatus } from "tsjs-xpx-chain-sdk";
-import { WalletService } from "./wallet.service";
-import { TransactionsService } from "../../transactions/service/transactions.service";
 import { environment } from '../../../environments/environment';
 import { NodeService } from '../../servicesModule/services/node.service';
 import { SharedService } from './shared.service';
-import { TransactionsInterface } from '../../dashboard/services/transaction.interface';
+import { WalletService } from '../../wallet/services/wallet.service';
+import { TransactionsInterface, TransactionsService } from '../../transfer/services/transactions.service';
 
 
 @Injectable({
@@ -83,6 +82,8 @@ export class DataBridgeService {
               unconfirmed.unshift(element);
             }
           }
+
+
           this.transactionsService.setTransactionsUnConfirmed$(unconfirmed);
         }
       });
@@ -109,7 +110,6 @@ export class DataBridgeService {
         this.reconnect(this.connector);
       });
     }
-
   }
 
   /**
@@ -137,6 +137,7 @@ export class DataBridgeService {
   }
 
 
+
   /**
    * Get the confirmed transactions from the socket
    *
@@ -158,7 +159,6 @@ export class DataBridgeService {
           this.destroyUnconfirmedTransaction(transactionFormatter);
           this.transactionsService.setTransactionsConfirmed$(transactionPushed);
           audio.play();
-          // this.messageService.changeMessage('balanceChanged');
           this.transactionsService.validateTypeTransaction(incomingTransaction.type);
           // this.namespaceService.buildNamespaceStorage();
           // this.transactionsService.updateBalance();
@@ -182,6 +182,7 @@ export class DataBridgeService {
         'type': 'unconfirmed',
         'data': unconfirmedTransaction
       });
+
       this.transactionsService.getTransactionsUnConfirmed$().pipe(first()).subscribe(
         async transactionsUnconfirmed => {
           const transactionPushed = transactionsUnconfirmed.slice(0);
