@@ -1,157 +1,185 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormBuilder, Validators } from "@angular/forms";
 import { NetworkType } from 'tsjs-xpx-chain-sdk';
-import { SharedService, WalletService } from "../../../shared";
+import { FormGroup, FormBuilder, Validators, AbstractControl } from '@angular/forms';
+import { Router } from '@angular/router';
+import { SharedService, ConfigurationForm } from '../../../shared/services/shared.service';
+import { WalletService } from '../../services/wallet.service';
 import { ProximaxProvider } from '../../../shared/services/proximax.provider';
+import { AppConfig } from 'src/app/config/app.config';
 
 
 @Component({
   selector: 'app-import-wallet',
   templateUrl: './import-wallet.component.html',
-  styleUrls: ['./import-wallet.component.scss']
+  styleUrls: ['./import-wallet.component.css']
 })
 export class ImportWalletComponent implements OnInit {
 
-  nameModule = 'Import Wallet';
-  descriptionModule = '';
   importWalletForm: FormGroup;
-  walletIsCreated = false;
-  pvk: string;
-  address: string;
-  typeNetwork = [
-    {
-      'value': NetworkType.TEST_NET,
-      'label': 'TEST NET'
-    }
-  ];
-  walletName: string;
-  publicKey: string;
+  configurationForm: ConfigurationForm = {};
+  description = 'Restores your existing Proximax SiriusWallet, import a private key from another service, or create a new wallet right now!';
+  errorMatchPassword: string;
+  errorWalletExist: string;
+  isValid: boolean = false;
+  title = 'Import Wallet';
+  typeNetwork = [{
+    value: NetworkType.TEST_NET,
+    label: 'TEST NET'
+  }];
 
   constructor(
     private fb: FormBuilder,
     private sharedService: SharedService,
-    private _walletService: WalletService,
-    private proximaxProvider: ProximaxProvider
-  ) {
-  }
+    private walletService: WalletService,
+    private proximaxProvider: ProximaxProvider,
+    private router: Router
+  ) { }
 
   ngOnInit() {
-    this.importForm();
+    this.configurationForm = this.sharedService.configurationForm;
+    this.createFormImportWallet();
   }
 
-  /**
-   *
-   *
-   * @memberof ImportWalletComponent
-   */
-  importForm() {
+  createFormImportWallet() {
     this.importWalletForm = this.fb.group({
-      walletname: ['', [Validators.required, Validators.minLength(2), Validators.maxLength(30)]],
-      network: [NetworkType.TEST_NET, [Validators.required]],
-      passwords: this.fb.group({
-        password: ['', [Validators.required, Validators.minLength(8), Validators.maxLength(30)]],
-        confirm_password: ['', [Validators.required, Validators.minLength(8), Validators.maxLength(30)]],
-      }, { validator: this.sharedService.passwordConfirming }),
-      privateKey: ['', [Validators.required, Validators.minLength(64), Validators.maxLength(64), Validators.pattern('^(0x|0X)?[a-fA-F0-9]+$')]],
+      nameWallet: ['', [
+        Validators.required,
+        Validators.minLength(this.configurationForm.nameWallet.minLength),
+        Validators.maxLength(this.configurationForm.nameWallet.maxLength)
+      ]],
+      privateKey: ['', [
+        Validators.required,
+        Validators.minLength(this.configurationForm.privateKey.minLength),
+        Validators.maxLength(this.configurationForm.privateKey.maxLength),
+        Validators.pattern('^(0x|0X)?[a-fA-F0-9]+$')
+      ]],
+      network: [
+        NetworkType.TEST_NET, [Validators.required]
+      ],
+      passwords: this.fb.group(
+        {
+          password: [
+            '', [
+              Validators.required,
+              Validators.minLength(this.configurationForm.passwordWallet.minLength),
+              Validators.maxLength(this.configurationForm.passwordWallet.maxLength)
+            ]
+          ],
+          confirm_password: [
+            '',
+            [
+              Validators.required,
+              Validators.minLength(this.configurationForm.passwordWallet.minLength),
+              Validators.maxLength(this.configurationForm.passwordWallet.maxLength)
+            ]
+          ],
+        }, {
+          validator: this.sharedService.equalsPassword
+        }),
     });
   }
 
-  /**
-   * Import a simple wallet
-   *
-   * @memberof ImportWalletComponent
-   */
-  importSimpleWallet() {
-    if (this.importWalletForm.valid) {
-      const nameWallet = this.importWalletForm.get('walletname').value;
-      const walletsStorage = this._walletService.getWalletStorage();
-      //verify if name wallet isset
-      const myWallet = walletsStorage.find(function (element) {
-        return element.name === nameWallet;
-      });
-
-      //Wallet does not exist
-      if (myWallet === undefined) {
-        const password = this.proximaxProvider.createPassword(this.importWalletForm.controls.passwords.get('password').value);
-        const privateKey = this.importWalletForm.get('privateKey').value;
-        const network = this.importWalletForm.get('network').value;
-        const wallet = this.proximaxProvider.createAccountFromPrivateKey(nameWallet, password, privateKey, network);
-        const accounts = this._walletService.buildAccount(wallet.encryptedPrivateKey.encryptedKey, wallet.encryptedPrivateKey.iv, wallet.address['address'], wallet.network);
-        this.walletName = nameWallet;
-        this._walletService.setAccountWalletStorage(nameWallet, accounts);
-        this.address = wallet.address.pretty();
-        this.sharedService.showSuccess('Congratulations!', 'Your wallet has been imported successfully');
-        this.pvk = this.proximaxProvider.decryptPrivateKey(password, accounts.encrypted, accounts.iv).toUpperCase();
-        this.publicKey = this.proximaxProvider.getPublicAccountFromPrivateKey(this.pvk, network).publicKey;
-        this.walletIsCreated = true;
-        this.nameModule = 'Congratulations!';
-        this.descriptionModule = 'Your wallet has been imported successfully';
-      } else {
-        //Error of repeated Wallet
-        this.clearForm('walletname');
-        this.sharedService.showError('Attention!', 'This name is already in use, try another name');
-      }
-    } else if (this.importWalletForm.controls.passwords.get('password').valid &&
-      this.importWalletForm.controls.passwords.get('confirm_password').valid &&
-      this.importWalletForm.controls.passwords.getError('noMatch')) {
-      this.sharedService.showError('Attention!', `Password doesn't match`);
-      this.clearForm('password', 'passwords');
-      this.clearForm('confirm_password', 'passwords');
-    }
-
-  }
-
-  /**
-   * Function that gets errors from a form
-   *
-   * @param {*} param
-   * @param {*} name
-   * @returns
-   * @memberof ImportWalletComponent
-   */
-  getError(control, formControl?) {
-    if (formControl === undefined) {
-      if (this.importWalletForm.get(control).getError('required')) {
-        return `This field is required`;
-      } else if (this.importWalletForm.get(control).getError('minlength')) {
-        return `This field must contain minimum ${this.importWalletForm.get(control).getError('minlength').requiredLength} characters`;
-      } else if (this.importWalletForm.get(control).getError('maxlength')) {
-        return `This field must contain maximum ${this.importWalletForm.get(control).getError('maxlength').requiredLength} characters`;
-      } else if (this.importWalletForm.get(control).getError('pattern')) {
-        return `This field content characters not permited`;
-      }
-    } else {
-      if (this.importWalletForm.controls[formControl].get(control).getError('required')) {
-        return `This field is required`;
-      } else if (this.importWalletForm.controls[formControl].get(control).getError('minlength')) {
-        return `This field must contain minimum ${this.importWalletForm.controls[formControl].get(control).getError('minlength').requiredLength} characters`;
-      } else if (this.importWalletForm.controls[formControl].get(control).getError('maxlength')) {
-        return `This field must contain maximum ${this.importWalletForm.controls[formControl].get(control).getError('maxlength').requiredLength} characters`;
-      } else if (this.importWalletForm.controls[formControl].getError('noMatch')) {
-        return `Password doesn't match`;
-      } else if (this.importWalletForm.controls[formControl].getError('pattern')) {
-        return `This field content characters not permited`;
-      }
-    }
-  }
-
-  /**
-   * Clear form
-   *
-   * @memberof ImportWalletComponent
-   */
-  clearForm(custom?, formControl?) {
-    if (custom !== undefined) {
-      if (formControl !== undefined) {
-        this.importWalletForm.controls[formControl].get(custom).reset();
+  clearForm(nameInput: string = '', nameControl: string = '') {
+    if (nameInput !== '') {
+      if (nameControl !== '') {
+        this.importWalletForm.controls[nameControl].get(nameInput).reset();
         return;
       }
-      this.importWalletForm.get(custom).reset();
+
+      this.importWalletForm.get(nameInput).reset();
       return;
     }
+
     this.importWalletForm.reset();
     this.importWalletForm.get('network').setValue(NetworkType.TEST_NET);
     return;
   }
-}
 
+  importSimpleWallet() {
+    if (this.importWalletForm.valid && this.isValid) {
+      //verify if name wallet isset
+      const existWallet = this.walletService.getWalletStorage().find(
+        (element: any) => {
+          return element.name === this.importWalletForm.get('nameWallet').value;
+        }
+      );
+
+      //Wallet does not exist
+      if (existWallet === undefined) {
+        const nameWallet = this.importWalletForm.get('nameWallet').value;
+        const network = this.importWalletForm.get('network').value;
+        const privateKey = this.importWalletForm.get('privateKey').value;
+        const password = this.proximaxProvider.createPassword(this.importWalletForm.controls.passwords.get('password').value);
+        const wallet = this.proximaxProvider.createAccountFromPrivateKey(nameWallet, password, privateKey, network);
+        const dataAccount = this.walletService.buildAccount(
+          wallet.encryptedPrivateKey.encryptedKey,
+          wallet.encryptedPrivateKey.iv,
+          wallet.address['address'],
+          wallet.network
+        );
+
+        this.clearForm();
+        this.walletService.saveDataWalletCreated({
+          name: nameWallet,
+          algo: password,
+          network: wallet.network
+        }, dataAccount, wallet);
+        this.walletService.saveAccountStorage(nameWallet, dataAccount);
+        this.router.navigate([`/${AppConfig.routes.walletCreated}`]);
+      } else {
+        //Error of repeated Wallet
+        this.clearForm('nameWallet');
+        this.sharedService.showError('', 'This name is already in use, try another name');
+      }
+    }
+  }
+
+  validateInput(nameInput: string = '', nameControl: string = '', nameValidation: string = '') {
+    let validation: AbstractControl = null;
+    if (nameInput !== '' && nameControl !== '') {
+      validation = this.importWalletForm.controls[nameControl].get(nameInput);
+    } else if (nameInput === '' && nameControl !== '' && nameValidation !== '') {
+      validation = this.importWalletForm.controls[nameControl].getError(nameValidation);
+    } else if (nameInput !== '') {
+      validation = this.importWalletForm.get(nameInput);
+    }
+    return validation;
+  }
+
+
+  validateMatchPassword() {
+    if (this.validateInput('password', 'passwords').valid &&
+      this.validateInput('confirm_password', 'passwords').valid &&
+      this.validateInput('', 'passwords', 'noMatch') &&
+      (this.validateInput('password', 'passwords').dirty || this.validateInput('password', 'passwords').touched) &&
+      (this.validateInput('password', 'passwords').dirty || this.validateInput('password', 'passwords').touched)) {
+      this.errorMatchPassword = '-invalid';
+      return true;
+    }
+
+    this.errorMatchPassword = '';
+    return false;
+  }
+
+
+  validateNameWallet() {
+    if (this.importWalletForm.get('nameWallet').valid) {
+      const existWallet = this.walletService.getWalletStorage().find(
+        (element: any) => {
+          return element.name === this.importWalletForm.get('nameWallet').value;
+        }
+      );
+
+      if (existWallet !== undefined) {
+        this.isValid = false;
+        this.errorWalletExist = '-invalid';
+        return true;
+      }else {
+        this.isValid = true;
+        this.errorWalletExist = '';
+        return false;
+      }
+    }
+  }
+
+}
